+++ src/cpp/fastrtps_deprecated/types/DynamicType.cpp
@@ -34,7 +34,8 @@ DynamicType::DynamicType()
 {
 }
 
-DynamicType::DynamicType(const TypeDescriptor* descriptor)
+DynamicType::DynamicType(
+        const TypeDescriptor* descriptor)
     : is_key_defined_(false)
 {
     descriptor_ = new TypeDescriptor(descriptor);
@@ -53,14 +54,15 @@ DynamicType::DynamicType(const TypeDescriptor* descriptor)
     if (kind_ == TK_ALIAS)
     {
         for (auto it = descriptor_->get_base_type()->member_by_id_.begin();
-            it != descriptor_->get_base_type()->member_by_id_.end(); ++it)
+                it != descriptor_->get_base_type()->member_by_id_.end(); ++it)
         {
             member_by_name_.insert(std::make_pair(it->second->get_name(), it->second));
         }
     }
 }
 
-DynamicType::DynamicType(const DynamicTypeBuilder* other)
+DynamicType::DynamicType(
+        const DynamicTypeBuilder* other)
     : descriptor_(nullptr)
     , name_("")
     , kind_(TK_NONE)
@@ -74,7 +76,8 @@ DynamicType::~DynamicType()
     clear();
 }
 
-ReturnCode_t DynamicType::apply_annotation(AnnotationDescriptor& descriptor)
+ReturnCode_t DynamicType::apply_annotation(
+        AnnotationDescriptor& descriptor)
 {
     if (descriptor.is_consistent())
     {
@@ -175,7 +178,8 @@ void DynamicType::clear()
     member_by_name_.clear();
 }
 
-ReturnCode_t DynamicType::copy_from_builder(const DynamicTypeBuilder* other)
+ReturnCode_t DynamicType::copy_from_builder(
+        const DynamicTypeBuilder* other)
 {
     if (other != nullptr)
     {
@@ -203,7 +207,8 @@ ReturnCode_t DynamicType::copy_from_builder(const DynamicTypeBuilder* other)
     }
 }
 
-bool DynamicType::exists_member_by_name(const std::string& name) const
+bool DynamicType::exists_member_by_name(
+        const std::string& name) const
 {
     if (descriptor_->get_base_type() != nullptr)
     {
@@ -215,7 +220,8 @@ bool DynamicType::exists_member_by_name(const std::string& name) const
     return member_by_name_.find(name) != member_by_name_.end();
 }
 
-ReturnCode_t DynamicType::get_descriptor(TypeDescriptor* descriptor) const
+ReturnCode_t DynamicType::get_descriptor(
+        TypeDescriptor* descriptor) const
 {
     if (descriptor != nullptr)
     {
@@ -251,10 +257,12 @@ bool DynamicType::key_annotation() const
     return false;
 }
 
-bool DynamicType::equals(const DynamicType* other) const
+bool DynamicType::equals(
+        const DynamicType* other) const
 {
     if (other != nullptr && descriptor_->annotation_.size() == other->descriptor_->annotation_.size() &&
-        member_by_id_.size() == other->member_by_id_.size() && member_by_name_.size() == other->member_by_name_.size())
+            member_by_id_.size() == other->member_by_id_.size() &&
+            member_by_name_.size() == other->member_by_name_.size())
     {
         // Check the annotation list
         for (auto it = descriptor_->annotation_.begin(),
@@ -337,7 +345,8 @@ ReturnCode_t DynamicType::get_member_by_name(
     }
 }
 
-ReturnCode_t DynamicType::get_all_members_by_name(std::map<std::string, DynamicTypeMember*>& members)
+ReturnCode_t DynamicType::get_all_members_by_name(
+        std::map<std::string, DynamicTypeMember*>& members)
 {
     members = member_by_name_;
     return ReturnCode_t::RETCODE_OK;
@@ -360,7 +369,8 @@ ReturnCode_t DynamicType::get_member(
     }
 }
 
-ReturnCode_t DynamicType::get_all_members(std::map<MemberId, DynamicTypeMember*>& members)
+ReturnCode_t DynamicType::get_all_members(
+        std::map<MemberId, DynamicTypeMember*>& members)
 {
     members = member_by_id_;
     return ReturnCode_t::RETCODE_OK;
@@ -396,13 +406,14 @@ DynamicType_ptr DynamicType::get_base_type() const
     return DynamicType_ptr(nullptr);
 }
 
-uint32_t DynamicType::get_bounds(uint32_t index /*= 0*/) const
+uint32_t DynamicType::get_bounds(
+        uint32_t index /*= 0*/) const
 {
     if (descriptor_ != nullptr)
     {
         return descriptor_->get_bounds(index);
     }
-    return static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED);
+    return dds::core::LENGTH_UNLIMITED;
 }
 
 uint32_t DynamicType::get_bounds_size() const
@@ -447,19 +458,20 @@ uint32_t DynamicType::get_total_bounds() const
     {
         return descriptor_->get_total_bounds();
     }
-    return static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED);
+    return dds::core::LENGTH_UNLIMITED;
 }
 
 bool DynamicType::has_children() const
 {
     return kind_ == TK_ANNOTATION || kind_ == TK_ARRAY || kind_ == TK_MAP || kind_ == TK_SEQUENCE
-        || kind_ == TK_STRUCTURE || kind_ == TK_UNION || kind_ == TK_BITSET;
+           || kind_ == TK_STRUCTURE || kind_ == TK_UNION || kind_ == TK_BITSET;
 }
 
 bool DynamicType::is_complex_kind() const
 {
     return kind_ == TK_ANNOTATION || kind_ == TK_ARRAY || kind_ == TK_BITMASK || kind_ == TK_ENUM
-        || kind_ == TK_MAP || kind_ == TK_SEQUENCE || kind_ == TK_STRUCTURE || kind_ == TK_UNION || kind_ == TK_BITSET;
+           || kind_ == TK_MAP || kind_ == TK_SEQUENCE || kind_ == TK_STRUCTURE || kind_ == TK_UNION ||
+           kind_ == TK_BITSET;
 }
 
 bool DynamicType::is_consistent() const
@@ -474,12 +486,13 @@ bool DynamicType::is_discriminator_type() const
         return descriptor_->get_base_type()->is_discriminator_type();
     }
     return kind_ == TK_BOOLEAN || kind_ == TK_BYTE || kind_ == TK_INT16 || kind_ == TK_INT32 ||
-        kind_ == TK_INT64 || kind_ == TK_UINT16 || kind_ == TK_UINT32 || kind_ == TK_UINT64 ||
-        kind_ == TK_FLOAT32 || kind_ == TK_FLOAT64 || kind_ == TK_FLOAT128 || kind_ == TK_CHAR8 ||
-        kind_ == TK_CHAR16 || kind_ == TK_STRING8 || kind_ == TK_STRING16 || kind_ == TK_ENUM || kind_ == TK_BITMASK;
+           kind_ == TK_INT64 || kind_ == TK_UINT16 || kind_ == TK_UINT32 || kind_ == TK_UINT64 ||
+           kind_ == TK_FLOAT32 || kind_ == TK_FLOAT64 || kind_ == TK_FLOAT128 || kind_ == TK_CHAR8 ||
+           kind_ == TK_CHAR16 || kind_ == TK_STRING8 || kind_ == TK_STRING16 || kind_ == TK_ENUM || kind_ == TK_BITMASK;
 }
 
-void DynamicType::set_name(const std::string& name)
+void DynamicType::set_name(
+        const std::string& name)
 {
     if (descriptor_ != nullptr)
     {
@@ -490,7 +503,7 @@ void DynamicType::set_name(const std::string& name)
 
 size_t DynamicType::get_size() const
 {
-    switch(kind_)
+    switch (kind_)
     {
         case TK_BOOLEAN: case TK_BYTE: case TK_CHAR8: return 1;
         case TK_INT16: case TK_UINT16: case TK_CHAR16:  return 2;
